Added a try/catch for std::out_of_range on frame_loggers
Consistently as the program is exiting, a stack.get_id() is returning a
std::thread::id that is not in the frame_loggers std::unordered_map. I'm
not doing anything but printing a message, which is not the right way to
handle this, but as the program is literally exiting, it seems livable
for the time being.