Changed the sytanx to separate the class defs from the object decls.
I found it a little confusing, it looked like the old C-style typedef
Struct... etc. sytnax and made it look like it was trying to declare
a static class, which is not allowed in C++11 and later. So the class
defs are now just that, and the static object declarations are separate
statements.