_m_plugin_id from topic_subscription initialization list to ctor body
Clang -Werror wouldn't compile because it said _m_plugin_id was an
unused variable. At first glance, this appeared to be correct because
the only place it was in the code was in the private declarations for
the topic_subscription class and in the class constructor initialization
list. However, "plugin ids", the actual idea, is used all throughout the
code, they're just usually accessed through a std::list of
topic_subscriptions instead of directly.

Deleting it, on the presumption that clang was correct, just led to the
same issue with _m_plugin_name, which, as a concept, is also used
throughout the code.

Changing from the initialization list to an assignment in
the constructor body made this error go away.