Removed comment reference to a "test.sh"
The top of file contains a large comment describing the timer
infrastructure. It references a test.sh which can be used to determine
the timing infrastructure overhead. This file exists only on the
project-scheduling branch. (Reminder - this PR was created from bits and
pieces of the project-scheduling and issue-209-cpu-timer branches, as
well as PR #221.) It is a BASH script which uses bazel. However, I can't locate
the actual tests. E.g. the first bazel command is

bazel run //include:cpu_timer_minimal_test \ # etc

ripgrep returns nothing for "cpu_timer_minimal_test" and find returns
there are no files with that file name.