Added [[maybe_unused]] attribute to MessageCallback
clang++-10 was not compiling when using configs/native-gl.yaml because
it claimed common/shader_util.hpp MessageCallback() was unused. This is
confusing because gldemo and timewarp_gl both appear to use this as
parameters for glDebugMessageCallback().

One possible fix would be to ad -Wno-unused-function as a compiler flag,
but that affects all compiled code. Using the attribute limits this to
the specific instance clang++-10 was angry about.