Put only spdlog call not inside #ifndef NDEBUG/#endif, inside.
There are 7 calls to spdlog::get() in common/switchboard.hpp. Six of
these were already inside of #ifndef NDEBUG/#endif blocks:
3 in topic_subscription::thread_on_start()
1 in the reader class constructor
2 in  switchboard::try_register_topic()

The remaning call, which was not inside of a preprocessor block was in
the topic_buffer class constructor.

This one particular call caused a crash when using either the open_vins and/or
debugview plugin's compiled with the opt profile.

This seems like a reasonable fix, considering that the other calls in
switchboard.hpp are inside preprocessing blocks.